backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
google-cloud-firestore==2.21.0
google-cloud-aiplatform==1.93.0
google-genai==1.27.0
